+++ CHANGELOG.md
@@ -13,6 +13,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+## [1.0.6] - 2026-09-08
+
+### Added
+
+- New folder and Delete in both panes; delete confirms and sends selected files or folders to MEGA/PikPak trash (#11)
+
+### Changed
+
+### Fixed
+
+- File pane loading overlay stays on top and the list cannot scroll underneath while Loading is shown
+
 ## [1.0.5] - 2026-09-07
 
 ### Added
@@ -108,7 +120,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-[Unreleased]: https://github.com/spellblade/Cloud-Copy/compare/v1.0.5...HEAD
+[Unreleased]: https://github.com/spellblade/Cloud-Copy/compare/v1.0.6...HEAD
+[1.0.6]: https://github.com/spellblade/Cloud-Copy/compare/v1.0.5...v1.0.6
 [1.0.5]: https://github.com/spellblade/Cloud-Copy/compare/v1.0.4...v1.0.5
 [1.0.4]: https://github.com/spellblade/Cloud-Copy/compare/v1.0.3...v1.0.4
 [1.0.3]: https://github.com/spellblade/Cloud-Copy/compare/v1.0.2...v1.0.3